+++ cucumber.json
@@ -3,12 +3,13 @@
     "formatOptions": {
       "snippetInterface": "async-await"
     },
-    "paths": ["src/test/features/saucedemo_login.feature"],
+    "paths": ["src/test/features/PimTrading/pimTrading_login.feature"],
     "dryRun": false,
     "require": [
-      "src/test/steps/free.ts",
       "src/hooks/hooks.ts",
-      "src/test/steps/add_item.ts"
+      "src/test/steps/fe/free.ts",
+      "src/test/steps/fe/workspace.ts",
+      "src/test/steps/PimTrading/trading_catalog.ts"
     ],
     "requireModule": ["ts-node/register"],
     "format": [
