+++ package.json
@@ -11,6 +11,8 @@
   "license": "ISC",
   "devDependencies": {
     "@playwright/test": "^1.37.1",
+    "chai": "^4.3.8",
+    "pg": "^8.11.3",
     "ts-node": "^10.9.1"
   },
   "dependencies": {
