+++ src/hooks/hooks.ts
@@ -1,7 +1,7 @@
 import { BeforeAll, AfterAll, setDefaultTimeout } from "@cucumber/cucumber";
 import { chromium, Browser, Page, webkit } from "@playwright/test";
 import { pageFixture } from "./pageFixture";
-import { defineParams } from "../commons/Define";
+import { defineParams } from "../commons/define";
 let page: Page;
 let browser: Browser;
 setDefaultTimeout(60 * 1000 * 2);
