--- src/pom/saucedemo/info.json
@@ -1,51 +1,0 @@
-{
-  "SauceDemo": {
-    "HomePage": {
-      "elements": {
-        "username_input": {
-          "selector": "#user-name",
-          "data_type": "string"
-        },
-        "password_input": {
-          "selector": "#password",
-          "data_type": "string"
-        },
-        "login_button": {
-          "selector": "#login-button",
-          "type": "button"
-        }
-      },
-      "url": "https://www.saucedemo.com"
-    },
-    "InventoryPage": {
-      "elements": {
-        "cart": {
-          "selector": ".shopping_cart_link",
-          "type": "button"
-        }
-      },
-      "url": "https://www.saucedemo.com/inventory.html"
-    }
-  },
-  "Ichiba": {
-    "LoginPage": {
-      "elements": {
-        "username_input": {
-          "selector": "#txt_email",
-          "data_type": "string"
-        },
-        "password_input": {
-          "selector": "#input_password"
-        },
-        "login_button": {
-          "selector": "#btn_submit"
-        }
-      },
-      "url": "https://test-org.ichiba.net/vi/"
-    },
-    "HomePage": {
-      "elements": {},
-      "url": "https://test-org.ichiba.net/vi/"
-    }
-  }
-}
