--- src/pom/saucedemo/login_pom.json
@@ -1,18 +1,0 @@
-{
-  "elements": {
-    "username_input": {
-      "locator": "#user-name",
-      "type": "string"
-    },
-    "password_input": {
-      "locator": "#password",
-      "type": "string"
-    },
-    "login_button": {
-      "locator": "#login-button",
-      "type": "button"
-    }
-  },
-  "url": "https://www.saucedemo.com",
-  "name": "SaucedemoHomepage"
-}
