+++ src/test/features/PimTrading/pimTrading_login.feature
@@ -5,10 +5,12 @@ Feature: Login Pim Trading
     When I enter input:username_input as "<username>"
     And I enter input:password_input as "<password>"
     And I click button:login_button
-    Then I navigate to PimTrading.HomePage successfully!
+    Then I navigate to PimTrading.WorkspacePage successfully!
+    When I choose workspace with workspace name "ws nam"
+    Then Browser send api get list subscribed source and get right response
+    Then I navigate to PimTrading.TradingCatalogPage with slug-name "ws-nam"
     Examples:
     |username|password|
-    |mongtoria@gmail.com|Cr@zyloop1|
     |namtvn@ichiba.vn|Cr@zyloop1|
